Remove gate check for mcp-qa
Until we have tests that can be executed in gate check scope,
we always fail job with exit 1. That's blcoked us from merges
some product related tests.
Remove gate check for mcp-qa project for now,and will back it as soon
as we find the way how to check mcp-qa

Change-Id: I9c40075e417dfcfb2e7336eaddae0b4484611917