Added new periodic pipeline to zuul and corresponding jobs.
This new pipeline is meant to trigger jobs automatically on a timer
(a'la cron). We want to trigger all jobs from zuul for 2 reasons:
1. Jenkins config is mostly volatile i.e. manual changes done to web-gui
wil dissapear after redeployment of Jenkins master.
2. Keeping all triggers in 1 place (zuul's layout.yaml ) is easier to manage.

A new job added will rebuild MCP Docker images and push them to public
registry with "BUILD_NUMBER" as a tag. This job will be triggered by
timer. Name of the job: mcp-build-images-build-number.

Second new job will be triggered on post-action after merge to OpenStack
components repositories and will push to registry with "latest" tag.
Name of the job: mcp-build-images-latest,

Change-Id: I71bf6e264372048209497fd2084ae9cf9b2dbd22