Add a test-build-stacklight job
This commit provides an alternative to the change
I93e560f608c968ed7f6fab8932e17c1e3a91e09f.

This commit reworks the test-build-{name} job template to
make it more generic, and actually handle the case where the
Git repo includes more than one build components.

It also adds a build for fuel-ccp-stacklight, relying on the
refactoring described above.

This is what the Hudson shell looks like for
test-build-stacklight job:

    <hudson.tasks.Shell>
      <command>
      virtualenv stacklightenv
      source stacklightenv/bin/activate
      pip install --upgrade pip
      pip install fuel-ccp/
      ccp --images-maintainer mirantis@mirantis.com \
          --images-namespace stacklightbuild --images-tag latest \
          --repositories-names fuel-ccp-debian-base,fuel-ccp-stacklight --builder-no-cache \
          --repositories-path containers/openstack \
          build -c elasticsearch grafana influxdb heka hindsight kibana snap
      deactivate
      </command>
    </hudson.tasks.Shell>

And for the test-build-nova:

    <hudson.tasks.Shell>
      <command>
      virtualenv novaenv
      source novaenv/bin/activate
      pip install --upgrade pip
      pip install fuel-ccp/
      ccp --images-maintainer mirantis@mirantis.com \
          --images-namespace novabuild --images-tag latest \
          --repositories-names fuel-ccp-debian-base,fuel-ccp-openstack-base,fuel-ccp-nova --builder-no-cache \
          --repositories-path containers/openstack \
          build -c nova
      deactivate
      </command>
    </hudson.tasks.Shell>

Right now any change to a repo will lead to rebuilding all
the build components of that repo. In the future we may want
to try rebuilding only the build components that are
impacted by the change.

Change-Id: I9eb5eac3cef86e59e69d7e94f748391904b76a0d