[WIP] Use DependentPipelineManager
This change is WIP and untested. The goal is to be able to
test fuel-ccp-* repos with fuel-ccp versions in the check
pipeline. Right now we always test fuel-ccp-* repos against
the master branch of fuel-ccp. This should eventually be
changed.

I am just opening this for discussion and reference.

References:
* http://blogs.rdoproject.org/7542/dive-into-zuul-gated-commit-system-2
* https://github.com/openstack-infra/project-config/blob/9bf9237965acff531b8c51a4e278c58824dde345/jenkins/jobs/macros.yaml#L29-L54

Change-Id: I0e43fb72dc9a9947aa151ea9ce83ec5d8f90a5e4