[WIP] Use DependentPipelineManager
This change is WIP and untested. The goal is to be able to
test fuel-ccp-* repos with fuel-ccp versions in the check
pipeline. Right now we always test fuel-ccp-* repos against
the master branch of fuel-ccp. This should eventually be
changed.

I am just opening this for discussion and reference.

Change-Id: I0e43fb72dc9a9947aa151ea9ce83ec5d8f90a5e4