Convert Docker registry cleanup job to manual mode
We no longer want to build periodically CCP Docker images with BUILD_NUMBER
as tag so we don't need automatic purge of old tags.
But sometimes it's useful to manually remove some unwanted tags.
This job is now meant to be run manually and accepts a parameter from
Jenkins web-ui which allows for selecting which tags should be deleted
(by regular expression).

Change-Id: I43b35e8cc8ba17fe22ba8014065938d32286c600