New CCP deployment and OpenStack test CI job
This job deploys CCP and runs simple, built-in OpenStack tests and makes
use of fuel-devops snaphosts to run faster skipping deployment of K8s cluster.
Kubernetes cluster is expected to be pre-deployed and snapshoted - if not
this job will ensure this so next builds of this job on the same slave will
re-use it.
It is not a replacement for "smoke" jobs but a workaround to provide any kind
of CI-based deployment and OpenStack tests until "smoke" gets more stable.

Change-Id: Ia508ddbb53fc7135e3046d7b2bc9ef3ab0b8b96b