Add new job for removing fuel-devops K8s snapshots
Removing snaphosts forces next builds of fuel-ccp deployment jobs
to recreate fuel-devops envs which is useful when we want to use
newer images of VMs.
This job is set to be automatically triggered on every node by job that
updates VM image on slaves.
Additionally this patch ads "remove-all-snapshots" job that can be used
manually to remove snaphosts on selected nodes.

Change-Id: I7f3d5e3531a9d950944c1ae202a4439ae5b67426