Dispatch build and deployment jobs to all nodes
It's turn on Least Load Plugin usage. In default dispatching rule
Jenkins is trying to run job on slave it was previously ran. So we can
have situation when we've free nodes and jobs are waiting for access to
busy node. This patch change dispatching rules for build and deployment jobs.

Change-Id: Ibcdd16925db07e97abf48a02596a7a6e753060f9